requirements.txt
langchain>=0.2
langgraph>=0.2
langchain-openai>=0.1
chromadb>=0.5
sentence-transformers>=3.0
pypdf>=4.2
unstructured>=0.15
scipy>=1.12
numpy>=1.26
pydantic>=2.7
typer>=0.12
fastapi>=0.111
uvicorn>=0.30
structlog>=24.1
python-dotenv>=1.0
pytest>=8.0
streamlit>=1.28
plotly>=5.17
pandas>=2.0
openpyxl>=3.1
python-docx>=0.8
reportlab>=4.0